// frontend/src/types/schema.ts
export type LocalizedStringSchema = Record<string, string>;

/** MangaResponse */
export interface MangaResponseSchema {
    result: "ok" | "error";
    /** @default "entity" */
    response: string;
    data: MangaSchema;
}

/** ChapterResponse */
export interface ChapterResponseSchema {
    result: "ok" | "error";
    /** @default "entity" */
    response: string;
    data: ChapterSchema;
}

/** Relationship */
export interface RelationshipSchema {
    /** @format uuid */
    id: string;
    type: string;
    /** Related Manga type, only present if you are on a Manga entity and a Manga relationship */
    related:
        | "monochrome"
        | "main_story"
        | "adapted_from"
        | "based_on"
        | "prequel"
        | "side_story"
        | "doujinshi"
        | "same_franchise"
        | "shared_universe"
        | "sequel"
        | "spin_off"
        | "alternate_story"
        | "alternate_version"
        | "preserialization"
        | "colored"
        | "serialization";
    /** If Reference Expansion is applied, contains objects attributes */
    attributes: object | null;
}

/** Chapter */
export interface ChapterSchema {
    /** @format uuid */
    id: string;
    type: "chapter";
    attributes: ChapterAttributesSchema;
    relationships: RelationshipSchema[];
}

/** Manga */
export interface MangaSchema {
    /** @format uuid */
    id: string;
    type: "manga";
    attributes: MangaAttributesSchema;
    relationships: RelationshipSchema[];
}
// ...
export interface MangaAttributesSchema {
    title: LocalizedStringSchema;
    altTitles: LocalizedStringSchema[];
    description: LocalizedStringSchema;
    isLocked: boolean;
    links: Record<string, string>;
    originalLanguage: string;
    lastVolume: string | null;
    lastChapter: string | null;
    publicationDemographic: "shounen" | "shoujo" | "josei" | "seinen" | null;
    status: "completed" | "ongoing" | "cancelled" | "hiatus";
    /** Year of release */
    year: number | null;
    contentRating: "safe" | "suggestive" | "erotica" | "pornographic";
    chapterNumbersResetOnNewVolume: boolean;
    availableTranslatedLanguages: string[];
    /** @format uuid */
    latestUploadedChapter: string;
    tags: TagSchema[];
    state: "draft" | "submitted" | "published" | "rejected";
    /** @min 1 */
    version: number;
    /** @format date-time */
    createdAt: Date;
    /** @format date-time */
    updatedAt: Date;
}
// ...
export interface CoverAttributesSchema {
    volume: string | null;
    fileName: string;
    description: string | null;
    locale: string | null;
    /** @min 1 */
    version: number;
    /** @format date-time */
    createdAt: Date;
    /** @format date-time */
    updatedAt: Date;
}
// ...
export interface TagSchema {
    /** @format uuid */
    id: string;
    type: "tag";
    attributes: TagAttributesSchema;
    relationships: RelationshipSchema[];
}